Add files via upload
Decentralized Used clothes marketplace ecosystem friendly on Connex Vechain Docs
With full Decentralized Network

Our wallet operates on the decentralized VechainThor network, avoiding the risks of centralization. Your assets are secure on the blockchain due to the absence of a single point of failure.

Connex serves as an "online" library, primarily used for dApp development and blockchain-related operations that can be performed online. For example, users can use Connex to create, sign and broadcast transactions to the VechainThor blockchain network.

Connex is the standard interface to connect decentralized applications (dApps) and users with the VechainThor blockchain. This documentation provides a comprehensive guide to using the Connex library, including installation, usage instructions, and details about its key features and methods.

It can be utilized through two distinct methodologies: as a Browser application or as a conventional Command Line Interface (CLI) application.
Whatever your Web3 innovation is, VeChain is the place to build it on.
Build your  future economy on the sustainable Web3  on VechainThor network.