Update index.ts
Decentralized used Clothes marketplace on Connex Vechain ecosystem.

VeChain (VET) is a blockchain-based smart contract platform that aims to improve supply chain management and product authentication. It was founded in 2015 and has gained significant attention in the blockchain and logistics communities.

Connex, on the other hand, is a subsidiary of VeChain, focused on developing a decentralized data exchange protocol for the Internet of Things (IoT). Connex aims to enable secure, decentralized data sharing and monetization for IoT devices, which is crucial for various industries like supply chain management, smart cities, and industrial automation.

VeChain and Connex work together to provide a comprehensive solution for supply chain management, product authentication, and data sharing, utilizing blockchain and IoT technologies.


https://github.com/Fesotam/connex/releases/tag/v2.1.0